feat: register agents router in app router
