--- src/trpc/middleware/routers/hello.ts
@@ -1,6 +1,0 @@
-import publicProcedure from '../../procedures/public';
-import { router } from '../../trpc';
-
-export const helloRouter = router({
-  hello: publicProcedure.query(() => 'Hello World'),
-});
